Refactor : WeatherXYChange.kt 삭제, 중기기온, 중기육상예보 폴더 구분, 변경 -> 그에 따라 파일명, 메소드명 변경
