+++ app/src/main/java/com/wannabeinseoul/seoulpublicservice/kma/midLandFcst/MidLandFcstApiService.kt
@@ -1,11 +1,11 @@
-package com.wannabeinseoul.seoulpublicservice.kma
+package com.wannabeinseoul.seoulpublicservice.kma.midLandFcst
 
 import com.wannabeinseoul.seoulpublicservice.BuildConfig
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
 
-interface KmaApiService {
+interface MidLandFcstApiService {
     @GET("MidFcstInfoService/getMidLandFcst")
     suspend fun getMidLandFcst(
         @Query("ServiceKey") serviceKey: String = BuildConfig.KMA_API_KEY,
