+++ app/src/main/java/com/wannabeinseoul/seoulpublicservice/kma/midLandFcst/MidLandFcstDto.kt
@@ -1,4 +1,4 @@
-package com.wannabeinseoul.seoulpublicservice.kma
+package com.wannabeinseoul.seoulpublicservice.kma.midLandFcst
 
 import com.google.gson.annotations.SerializedName
