+++ app/src/main/java/com/wannabeinseoul/seoulpublicservice/kma/midTemp/MidTempApiService.kt
@@ -1,11 +1,11 @@
-package com.wannabeinseoul.seoulpublicservice.kma.temperature
+package com.wannabeinseoul.seoulpublicservice.kma.midTemp
 
 import com.wannabeinseoul.seoulpublicservice.BuildConfig
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
 
-interface TempApiService {
+interface MidTempApiService {
     @GET("MidFcstInfoService/getMidTa")
     suspend fun getTemp(
         @Query("ServiceKey") serviceKey: String = BuildConfig.TEMP_API_KEY,
