+++ app/src/main/java/com/wannabeinseoul/seoulpublicservice/kma/midTemp/MidTempRepository.kt
@@ -1,4 +1,4 @@
-package com.wannabeinseoul.seoulpublicservice.kma.temperature
+package com.wannabeinseoul.seoulpublicservice.kma.midTemp
 
 import retrofit2.Response
 
@@ -13,7 +13,7 @@ interface TempRepository {
 }
 
 class TempRepositoryImpl (
-    private val tempApiService: TempApiService
+    private val midTempApiService: MidTempApiService
 ): TempRepository {
     override suspend fun getTemp(
         numOfRows: Int,
@@ -22,6 +22,6 @@ class TempRepositoryImpl (
         regId: String,
         tmFc: String
     ): Response<TemperatureDTO> {
-        return tempApiService.getTemp(numOfRows = numOfRows, pageNo = pageNo, dataType = dataType, regId = regId, tmFc = tmFc)
+        return midTempApiService.getTemp(numOfRows = numOfRows, pageNo = pageNo, dataType = dataType, regId = regId, tmFc = tmFc)
     }
 }
